Deploy InfluxDB in cluster mode
This change deploys InfluxDB on the 3 monitoring nodes and configures
HAProxy to forward requests only to the first active node. When the
first node fails, traffic will go to the second node (if active) and
eventually to the third node if the second node goes down too. Whenever
the first node comes back online, HAProxy will resume sending the
traffic to it. This is a poor's man HA solution that only ensures that
the operator can visualize the metrics even when the primary node is
unavailable but it doesn't account for data replication at all.