Enable Heka sensu output
This patch activates the Heka sensu output for all nodes because all
nodes are running Heka processes and we need to activate it for
registration and also to use the watchdog for monitoring the Heka
pipeline. It also sets the handlers that will be used to manage
alarms that are enabled with and without notifications.

Change-Id: I14ab2b6c2b506a50dc90dfedc55037c62d508a01